+++ src/lib.rs
@@ -18,8 +18,6 @@ pub mod collect_into_bit_set;
 
 pub type SetElement = u32;
 
-use crate::bit_set_n::{BitSet8, BitSet16, BitSet32, BitSet64, BitSet128};
-
 pub mod prelude {
     pub use crate::bit_set_array::BitSetArray;
     pub use crate::bit_set_n::*;
@@ -27,11 +25,13 @@ pub mod prelude {
     #[cfg(any(test, feature = "std"))]
     pub use crate::bit_set_vec::BitSetVec;
     pub use crate::finite::FiniteBitSet;
+    pub use crate::SetElement;
+    pub use crate::collect_into_bit_set::CollectIntoBitSet;
 }
 
 #[inline]
-pub(crate) fn mutate_inner<R>(inner: &mut u64, f: impl FnOnce(&mut BitSet64) -> R) -> R {
-    let mut set = BitSet64::from_inner_const(*inner);
+pub(crate) fn mutate_inner<R>(inner: &mut u64, f: impl FnOnce(&mut prelude::BitSet64) -> R) -> R {
+    let mut set = prelude::BitSet64::from_inner_const(*inner);
     let result = f(&mut set);
     *inner = set.into_inner_const();
     result