+++ CHANGELOG.md
@@ -11,6 +11,11 @@ Possible header types:
 
 [crates.io]: https://crates.io/crates/const_sized_bit_set
 
+## v0.5.0 (2025-12-31)
+- Reorganized the prelude
+- Added `len_const` to `BitSet8`, `BitSet16` etc.
+- Added `IntoIter` to `BitSet8`, `BitSet16` etc.
+
 ## v0.4.0 (2025-12-18)
 
 ### Breaking Changes